Game checks for winner, but experiencing closure errors
Currently, the game is not capable of being reset, and the game alerts
that the winner is player 1 regardless of whether X or O has three
in a row. I think that this is a closure error, I need to reread the article
on closure and rewrite some of the code.